#17110 Move instantiation mock to platform/instantiaion layer
