+++ src/vs/platform/actions/test/common/actions.test.ts
@@ -6,11 +6,12 @@
 
 import assert = require('assert');
 import WinJS = require('vs/base/common/winjs.base');
-import { TestInstantiationService } from 'vs/test/utils/instantiationTestUtils';
+import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
 import { DeferredAction } from 'vs/platform/actions/common/actions';
 import Actions = require('vs/base/common/actions');
 import { AsyncDescriptor } from 'vs/platform/instantiation/common/descriptors';
 import { IEventService } from 'vs/platform/event/common/event';
+import { EventService } from 'vs/platform/event/common/eventService';
 
 export class TestAction extends Actions.Action {
 	private service;
@@ -35,7 +36,7 @@ suite('Platform actions', () => {
 	test('DeferredAction', (done) => {
 
 		let instantiationService: TestInstantiationService = new TestInstantiationService();
-		instantiationService.stub(IEventService);
+		instantiationService.stub(IEventService, EventService);
 
 		let action = new DeferredAction(
 			instantiationService,
