+++ src/vs/platform/extensionManagement/test/common/extensionEnablementService.test.ts
@@ -8,7 +8,7 @@ import * as sinon from 'sinon';
 import * as assert from 'assert';
 import { IExtensionManagementService, IExtensionEnablementService, DidUninstallExtensionEvent } from 'vs/platform/extensionManagement/common/extensionManagement';
 import { ExtensionEnablementService } from 'vs/platform/extensionManagement/common/extensionEnablementService';
-import { TestInstantiationService } from 'vs/test/utils/instantiationTestUtils';
+import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
 import { TestEnvironmentService, TestWorkspace } from 'vs/test/utils/servicesTestUtils';
 import { Emitter } from 'vs/base/common/event';
 import { StorageService, InMemoryLocalStorage } from 'vs/platform/storage/common/storageService';
