+++ src/vs/workbench/parts/search/test/browser/searchViewlet.test.ts
@@ -7,7 +7,7 @@
 import * as assert from 'assert';
 import uri from 'vs/base/common/uri';
 import { Match, FileMatch, SearchResult } from 'vs/workbench/parts/search/common/searchModel';
-import { TestInstantiationService } from 'vs/test/utils/instantiationTestUtils';
+import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
 import { SearchSorter, SearchDataSource } from 'vs/workbench/parts/search/browser/searchResultsView';
 import { IFileMatch, ILineMatch } from 'vs/platform/search/common/search';
 import { createMockModelService } from 'vs/test/utils/servicesTestUtils';
