+++ src/vs/workbench/parts/search/test/common/searchResult.test.ts
@@ -6,11 +6,11 @@
 
 import * as assert from 'assert';
 import * as sinon from 'sinon';
-import { TestInstantiationService } from 'vs/test/utils/instantiationTestUtils';
+import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
 import { Match, FileMatch, SearchResult, SearchModel } from 'vs/workbench/parts/search/common/searchModel';
 import URI from 'vs/base/common/uri';
 import { IFileMatch, ILineMatch } from 'vs/platform/search/common/search';
-import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
+import { ITelemetryService, NullTelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { Range } from 'vs/editor/common/core/range';
 import { createMockModelService } from 'vs/test/utils/servicesTestUtils';
 import { IModelService } from 'vs/editor/common/services/modelService';
@@ -22,8 +22,10 @@ suite('SearchResult', () => {
 
 	setup(() => {
 		instantiationService = new TestInstantiationService();
-		instantiationService.stub(ITelemetryService);
+		instantiationService.stub(ITelemetryService, NullTelemetryService);
 		instantiationService.stub(IModelService, createMockModelService(instantiationService));
+		instantiationService.stubPromise(IReplaceService, {});
+		instantiationService.stubPromise(IReplaceService, 'replace', null);
 	});
 
 	test('Line Match', function () {
