+++ src/vs/workbench/parts/terminal/test/electron-browser/terminalInstance.test.ts
@@ -7,12 +7,11 @@
 
 import * as assert from 'assert';
 import * as os from 'os';
-import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
 import { IMessageService } from 'vs/platform/message/common/message';
 import { IStringDictionary } from 'vs/base/common/collections';
 import { IWorkspace } from 'vs/platform/workspace/common/workspace';
 import { TerminalInstance } from 'vs/workbench/parts/terminal/electron-browser/terminalInstance';
-import { TestInstantiationService } from 'vs/test/utils/instantiationTestUtils';
+import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
 import { TestMessageService } from 'vs/test/utils/servicesTestUtils';
 
 suite('Workbench - TerminalInstance', () => {
@@ -21,7 +20,6 @@ suite('Workbench - TerminalInstance', () => {
 
 	setup(() => {
 		instantiationService = new TestInstantiationService();
-		instantiationService.stub(IKeybindingService);
 		instantiationService.stub(IMessageService, new TestMessageService());
 	});
